feat(InputView): 자동으로 생성할 다리 길이를 입력 받는다.
