+++ docs/README.md
@@ -39,8 +39,8 @@
 
 
 ## 입력
-- [ ] 자동으로 생성할 다리 길이를 입력 받는다.
-  - [ ] 올바른 값이 아니면 예외 처리한다.
+- [x] 자동으로 생성할 다리 길이를 입력 받는다.
+  - [x] 올바른 값이 아니면 예외 처리한다.
 - [ ] 이동할 칸을 입력 받는다.
   - [ ] 올바른 값이 아니면 예외 처리한다.
 - [ ] 게임 재시작/종료 여부를 입력 받는다.
